Move container helpers into the resource
